User - Product cross-functionality
Created functions where users can submit product reviews, and like or remove likes from specific products. Another function also gets the average rating for a specific product.

At this stage, classes have not yet been introduced. However, in real applications this user-product functionality would be methods inside the User and Product classes, for example a likeProduct() method in the User class, which interacts with a product instance, and a getAvgRating() method in the Product class.